pyproject.toml
[project]
name = "langchain-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.11.3",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langsmith>=0.7.14",
    "python-dotenv>=1.2.2",
]
